check if an addon that is not a platform exists

+++ modules/core-services/services/project/project.go
@@ -649,15 +649,16 @@ func (p *Project) Delete(projectID int64) (*model.Project, error) {
 		return nil, errors.Errorf(p.trans.Text(langCodes, "FailedGetProject")+"(%v)", err)
 	}
 
-	// TODO We need to turn this check on after adding the delete portal to the UI
-	// check if addon exists
-	// addOnListResp, err := p.bdl.ListAddonByProjectID(projectID, project.OrgID)
-	// if err != nil {
-	// 	return nil, err
-	// }
-	// if addOnListResp != nil && len(addOnListResp.Data) > 0 {
-	// 	return nil, errors.Errorf("failed to delete project(there exists addons)")
-	// }
+	// Check if an addon that is not a platform exists
+	addOnListResp, err := p.bdl.ListAddonByProjectID(projectID, project.OrgID)
+	if err != nil {
+		return nil, err
+	}
+	if addOnListResp != nil && len(addOnsFilterIn(addOnListResp.Data, func(addOn *apistructs.AddonFetchResponseData) bool {
+		return addOn.PlatformServiceType == 0
+	})) > 0 {
+		return nil, errors.Errorf("failed to delete project(there exists addons)")
+	}
 
 	if err = p.db.DeleteProject(projectID); err != nil {
 		return nil, errors.Errorf(p.trans.Text(langCodes, "FailedDeleteProject")+"(%v)", err)
@@ -673,6 +674,15 @@ func (p *Project) Delete(projectID int64) (*model.Project, error) {
 		logrus.Warnf("failed to delete members extra, (%v)", err)
 	}
 	return &project, nil
+}
+
+func addOnsFilterIn(addOns []apistructs.AddonFetchResponseData, fn func(addOn *apistructs.AddonFetchResponseData) bool) (newAddons []apistructs.AddonFetchResponseData) {
+	for i := range addOns {
+		if fn(&addOns[i]) {
+			newAddons = append(newAddons, addOns[i])
+		}
+	}
+	return
 }
 
 // Get 获取项目
